Add macOS launch script
This script can be used to run the compiled JAR application
(build with mvn install first) on macOS with appropriate
parameters, such as a dock icon and quit on window close.

It can also be used to create a ".app" application for easy
access to Digital. With Automator, create a new "Application"
project, and add the action "Run Shell Script" with this command:

cd <PATH_TO_DIGITAL_FOLDER>/distribution/macos && ./Digital.sh

Signed-off-by: Vadim Chichikalyuk <vadim@vadimtch.eu>